simplify the test
